requirements.txt
configparser == 5.0.2

pip == 21.0.1

psycopg2-binary == 2.8.6
